Modification of tolerance
The mass flow tolerance parameter in the code is incredibly low which
means that occasionally (like with the hot water radiant/convective
baseboard) some noise gets through and causes a max iteration error.
This resets the low flow to just less than one gram of water (or one
drop) per hour.  This eliminates a host of max iteration errors that
are pretty unnecessary.