Corrections and unit test
Corrections made based on previous commit problems are included in
this.  Previous commit cut-off the baseboard from running at all
because the new logical condition in ControlCompOutput was not specific
and basically shut things down when starting the interval halving
process rather than after it got underway. Also a unit test for the fix
was implemented.